+++ src/AiMeter/Views/AuthWindow.xaml.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Linq;
 using System.Windows;
+using System.Windows.Interop;
 using AiMeter.Interop;
 using AiMeter.Services;
 using Microsoft.Web.WebView2.Core;
@@ -14,7 +15,7 @@ public partial class AuthWindow : Window
     public AuthWindow(IClaudeSession session)
     {
         InitializeComponent();
-        DarkTitleBar.Apply(this);
+        SourceInitialized += (s, e) => DarkTitleBar.ApplyToHwnd(new WindowInteropHelper(this).Handle);
         _session = session;
         InitializeAsync();
     }